--- test/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Mocha</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="mocha.css" />
-  </head>
-  <body>
-    <div id="mocha"></div>
-    <script src="mocha.js"></script>
-    <script>mocha.setup('bdd')</script>
-    <script src="https://rawgit.com/github/fetch/master/fetch.js"></script>
-    <script src="http://chaijs.com/chai.js"></script>
-    <script src="./../src/index.js"></script>
-    <script src="tests.js"></script>
-    <script>
-      mocha.run();
-    </script>
-  </body>
-</html>
